Modify BVH to have types & change scene objects ownership

+++ src/core/aabb.h
@@ -64,8 +64,8 @@ class AABB {
   bool intersect(Ray& ray) const {
     Interval tmp_interval(ray.t_min, ray.t_max);
     for (int i = 0; i < 3; i++) {
-      double t0 = (axis(i).min - ray.origin[i]) / ray.direction[i];
-      double t1 = (axis(i).max - ray.origin[i]) / ray.direction[i];
+      float t0 = (axis(i).min - ray.origin[i]) / ray.direction[i];
+      float t1 = (axis(i).max - ray.origin[i]) / ray.direction[i];
       if (t0 > t1) std::swap(t0, t1);
 
       // check whether overlaps
